Update README.md
Indicators of Heart Disease
Heart Disease is one of the leading causes of death in the United States. Heart Disease has some established risk factors such as High Blood Pressure, High Cholesterol and Smoking. As per the Center for Disease Control (CDC), 47% of all Americans have one of the three risk factors. Some other factors that we have no control but have an impact on Heart Disease include genetics, environmental factors and aging. In this dataset I explore two other indicators, Diabetes and higher than normal BMI, and their impact on heart disease.

A branch of CDC called the Behavioral Risk Factor Surveillance System (BRFSS) condcts a telephonic health survey annually that collects state data from U.S. residents regarding their health related risk behaviors and chronic health conditions. The BRFSS was established in 1984 with surveys conducted in 15 states and since then it has expanded to all 50 states, District of Columbia adn three U.S. territories. Every year, the BRFSS conducts about 400,000 interviews making it the largest continuously conducted health survey system in the world.

The latest survey data conducted was conducted in the year 2020. The dataset that I have used to do the analysis here, I found on Kaggle and is a subset of the data that comes from this survey. This dataset has over 300,000 respondents or rows and about 20 variables or columns.

Physicians and health care providers that are directly involved in the health of the population are constantly looking for more indicators that may accurately predict heart disease. Studying a large dataset like this over many years will continue to establish more patterns that can be used to either reduce the incidence of heart disease or tailor medications to prevent heart disease.

Tech companies that are working on wearables that monitor heart rate, step count, sleep, and many more in the pipeline are an emerging group that are constantly looking for a large dataset with many key health indicators. On one hand, tech wearables are getting more advanced and COVID has expanded how much telemedicine can truly accomplish, continuous surveying from the BRFSS program and data like these can help improve the lifestyles of individuals to reduce the risk of heart disease and in turn improve the health of the population.